+++ NEWS.md
@@ -1,7 +1,8 @@
 # googlePolylines 0.5.0 (development)
 
-
-* methods conversion back to `sfc` and `sf`
+* Example converting to `sfc` and `sf`
+* `as.data.frame.sfencoded()` to convert back to `data.frame` class
+* `decode()` - decodes encoded polylines to lat/lon
 
 # googlePolylines 0.4.0